+++ pkg/resources/database/utils.go
@@ -328,7 +328,7 @@ func UpdateResource[T ResourceModelInterface](ctx context.Context, req resource.
 	}
 
 	planData.WaitForService(ctx, client, &resp.Diagnostics)
-	planData.UpdateResource(ctx, client, &diags)
+	planData.UpdateResource(ctx, client, &resp.Diagnostics)
 
 	if resp.Diagnostics.HasError() {
 		return
@@ -374,7 +374,7 @@ func DeleteResource[T ResourceModelInterface](ctx context.Context, req resource.
 		return
 	}
 
-	data.DeleteResource(ctx, client, &diags)
+	data.DeleteResource(ctx, client, &resp.Diagnostics)
 
 	if resp.Diagnostics.HasError() {
 		return